% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{AvgCorr}
\alias{AvgCorr}
\title{Average Correlation Estimate.}
\usage{
AvgCorr(P1, P2, eps = 0.001)
}
\arguments{
\item{P1}{Fist p-value.}

\item{P2}{Second p-value.}

\item{eps}{Force correlation estimate to fall in the interval (eps,1-eps);}
}
\description{
Estimate correlation using the average of qnorm(p1)*qnorm(p2) across loci, where
(p1,p2) are p-values obtained via two different association tests.
}
