% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{BAT}
\alias{BAT}
\title{Basic Association Test}
\usage{
BAT(y, G, X, S)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Snp by obs genotype matrix.}

\item{X}{Obs by feature covariate matrix.}

\item{S}{Obs by feature structure matrix.}
}
\description{
Regression of phenotype on genotype, covariates, and structure.
}
