% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\name{BAT.WaldTest}
\alias{BAT.WaldTest}
\title{Basic Association Wald Test}
\usage{
BAT.WaldTest(y, G, X)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Genotype matrix with observations as rows, SNPs as columns.}

\item{X}{Model matrix of covariates.}
}
\value{
Numeric matrix, with 1 row per SNP, containing these columns:
\itemize{
  \item "score", the score statistic.
  \item "se", its standard error.
  \item "z", the Z statistic.
  \item "p", the p-value. 
}
}
\description{
Basic Association Wald Test
}
