% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OINT.R
\name{CauchyToP}
\alias{CauchyToP}
\title{Convert Cauchy Random Variable to P}
\usage{
CauchyToP(z)
}
\arguments{
\item{z}{Numeric Cauchy random variable.}
}
\value{
Numeric p-value.
}
\description{
Convert Cauchy Random Variable to P
}
