% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{FIINT}
\alias{FIINT}
\title{Fully Indirect-INT}
\usage{
FIINT(y, G, X, S, c = 3/8)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Snp by obs genotype matrix.}

\item{X}{Obs by feature covariate matrix.}

\item{S}{Obs by feature structure matrix.}

\item{c}{Offset applied during rank-normalization.}
}
\description{
Two-stage regression. Regresses phenotype on covariates and structure
to obtain residuals. Regresses transformed residuals on genotype.
}
