% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{Geno}
\alias{Geno}
\title{Simulated Genotypes}
\format{An integer matrix with 100 rows and 1000 columns
\describe{ 
   \item{s1-s1000}{s[i] is an integer vector of minor allele counts for the ith subject.}
}}
\usage{
Geno
}
\description{
Simulated genotypes for 1000 subjects at 100 loci. Genotypes were simulated 
with use of hapgen2, and are based on the haplotype structure of human 
chromosome one in the CEU population of the 1000 Genomes Project.
}
\keyword{datasets}
