% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IINT.R
\name{IINT.ScoreTest}
\alias{IINT.ScoreTest}
\title{Basic Association Score Test}
\usage{
IINT.ScoreTest(y, G, X, k)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Genotype matrix with observations as rows, SNPs as columns.}

\item{X}{Model matrix of covariates.}

\item{k}{Offset applied during rank-normalization.}
}
\value{
Numeric matrix, with 1 row per SNP, containing these columns:
\itemize{
  \item "score", the score statistic.
  \item "se", its standard error.
  \item "z", the Z statistic.
  \item "p", the p-value. 
}
}
\description{
Basic Association Score Test
}
