% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{missFilter}
\alias{missFilter}
\title{Missingness Filter}
\usage{
missFilter(y, G, X, S)
}
\arguments{
\item{y}{Numeric phenotype vector}

\item{G}{Snp by obs genotype matrix.}

\item{X}{Obs by feature covariate matrix.}

\item{S}{Obs by feature structure matrix.}
}
\description{
Function to adjust for missing data.
}
