% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{olsB}
\alias{olsB}
\title{Ordinary Least Squares Coeefficient}
\usage{
olsB(A, Y)
}
\arguments{
\item{A}{Numeric matrix}

\item{Y}{Numeric matrix}
}
\description{
Calculate the OLS coefficient \eqn{(A'A)^{-1}A'Y}.
}
