% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{Omni}
\alias{Omni}
\title{Omnibus P-value}
\usage{
Omni(Q)
}
\arguments{
\item{Q}{Numeric vector formatted as (p1,p2,rho), where p1 and p2 are 
estimated p-values, and rho is their correlation.}
}
\description{
Calculate p-value for omnibus statistic.
}
