% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\name{PartitionData}
\alias{PartitionData}
\title{Partition Data}
\usage{
PartitionData(e, g, X)
}
\arguments{
\item{e}{Numeric residual vector.}

\item{g}{Genotype vector.}

\item{X}{Model matrix of covariates.}
}
\value{
List containing:
\itemize{
  \item "g_obs", observed genotype vector.
  \item "X_obs", covariates for subjects with observed genotypes.
  \item "X_mis", covariates for subjects with missing genotypes. 
  \item "e_obs", residuals for subjects with observed genotypes.
}
}
\description{
Partition y and X according to the missingness pattern of g.
}
