% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{PIINT}
\alias{PIINT}
\title{Partially Indirect-INT}
\usage{
PIINT(y, G, X, S, M = T, c = 3/8)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Snp by obs genotype matrix.}

\item{X}{Obs by feature covariate matrix.}

\item{S}{Obs by feature structure matrix.}

\item{M}{Apply missingness filter?}

\item{c}{Offset applied during rank-normalization.}
}
\description{
Two-stage regression. Regresses phenotype on covariates to obtain 
residuals. Regresses transformed residuals on genotype and structure.
}
