% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OINT.R
\name{PtoCauchy}
\alias{PtoCauchy}
\title{Convert P-value to Cauchy Random}
\usage{
PtoCauchy(p)
}
\arguments{
\item{p}{Numeric p-value.}
}
\value{
Numeric Cauchy random variable.
}
\description{
Convert P-value to Cauchy Random
}
