% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{rankNormal}
\alias{rankNormal}
\title{Rank-Normalize}
\usage{
rankNormal(u, c = 3/8)
}
\arguments{
\item{u}{Numeric vector.}

\item{c}{Offset. Defaults to (3/8), correspond to the Blom transform.}
}
\value{
Numeric vector.
}
\description{
Applies the rank-normal transform to a numeric vector.
}
