% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{RNOmni}
\alias{RNOmni}
\title{Rank-Normal Omnibus Test}
\usage{
RNOmni(y, G, X, S, method = "AvgCorr", M = T, c = 3/8, B = 100,
  rho = F, parallel = F)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Snp by obs genotype matrix.}

\item{X}{Obs by feature covariate matrix.}

\item{S}{Obs by feature structure matrix.}

\item{method}{Method used to estimate correlation for the omnibus test,
either "AvgCorr" or "Bootstrap".}

\item{M}{Apply missingness filter?}

\item{c}{Offset applied during rank-normalization.}

\item{B}{Bootstrap samples for correlation estimation.}

\item{rho}{Logical indicating whether to return the correlation parameter
estimated during omnibus calculation. Defaults to FALSE.}

\item{parallel}{Run bootstraps in parallel? Must register parallel backend first.}
}
\description{
Omnibus association test that estimates p-values by direct and partially 
indirect inverse normal transformation (INT), then synthesizes the results
into a single test statistic, based on whichever approach provides more
evidence against null hypothesis.
}
