% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{S}
\alias{S}
\title{Population Structure Adjustments}
\format{A numeric matrix with 1000 rows and 2 columns.
\describe{
   \item{pc1}{The first principal component.}
   \item{pc2}{The second principal component.}
}}
\usage{
S
}
\description{
First two principal components of the genetic relatedness matrix. Note that
principal components were calculated using genotype at more loci than are
provided in the example data set.
}
\keyword{datasets}
