% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleParam}
\alias{scaleParam}
\title{Scale Parameter}
\usage{
scaleParam(y, Q, df)
}
\arguments{
\item{y}{Outcome}

\item{Q}{Error projection}

\item{df}{Degrees of freedom}
}
\description{
Scale Parameter
}
