% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreStat}
\alias{scoreStat}
\title{Score Statistic}
\usage{
scoreStat(y, Q, g, s2)
}
\arguments{
\item{y}{Outcome}

\item{Q}{Error projection}

\item{g}{Genotype vector}

\item{s2}{Scale parameter}
}
\description{
Score Statistic
}
