% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{X}
\alias{X}
\title{Simulated Covariates}
\format{A numeric matrix with 1000 rows and 2 columns
\describe{
   \item{Age}{Age.}
,,,\item{Sex}{Sex.}
}}
\usage{
X
}
\description{
Age and sex simulated for 1000 subjects. Age was drawn from a gamma
distribution with mean 50 and variance 10. Sex was drawn from a Bernoulli
distribuiton with expectation 1/2.
}
\keyword{datasets}
