% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{Y}
\alias{Y}
\title{Simulated Phenotypes}
\format{A numeric matrix with 1000 rows and 2 columns.
\describe{
   \item{YN}{Normal phenotype.}
   \item{YT3}{T3 phenotype.}
}}
\usage{
Y
}
\description{
Two independent phenotypes simulated under the null hypothesis of no 
genotypic effect. A subject specific mean was calculated based on
age, sex, pc1, and pc2. The normal phenotype was generated by adding N(0,1)
deviates to the subject means. The T3 phenotype was generated by adding \eqn{t_{3}/sqrt{3}} deviates
to the subject means.
}
\keyword{datasets}
