% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{AvgCorr}
\alias{AvgCorr}
\title{Average Correlation Estimate.}
\usage{
AvgCorr(P1, P2, eps = 0.001)
}
\arguments{
\item{P1}{Fist p-value.}

\item{P2}{Second p-value.}

\item{eps}{Force correlation estimate to fall in (eps,1-eps);}
}
\description{
Estimate correlation using the average of z1*z2 across loci.
}
