% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{AvgCorr}
\alias{AvgCorr}
\title{Average Correlation Estimate.}
\usage{
AvgCorr(p1, p2, a = 0.001)
}
\arguments{
\item{p1}{Fist p-value.}

\item{p2}{Second p-value.}

\item{a}{Force correlation estimate to fall in the interval (a,1-a);}
}
\description{
Estimate correlation using the average of qnorm(p1)*qnorm(p2) across loci, where
(p1,p2) are p-values obtained via two different association tests.
}
