% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputCheck.R
\name{BasicInputChecks}
\alias{BasicInputChecks}
\title{Basic Input Checks}
\usage{
BasicInputChecks(y, G, X)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Genotype matrix with observations as rows, SNPs as columns.}

\item{X}{Covariate matrix.}
}
\value{
None.
}
\description{
Stops evaluation if inputs are improperly formatted.
}
