% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{BootCorr}
\alias{BootCorr}
\title{Bootstrap Correlation Estimate.}
\usage{
BootCorr(y, G, X, S, B = 100, parallel)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Snp by obs genotype matrix.}

\item{X}{Obs by feature covariate matrix.}

\item{S}{Obs by feature structure matrix.}

\item{B}{Bootstrap samples for correlation estimation.}

\item{parallel}{Run bootstraps in parallel? Must register parallel backend first.}
}
\description{
Use bootstrap to estimate correlation among Z statistics
whose maximum is taken in the omnibus test.
}
