% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankNormal.R
\name{DINT}
\alias{DINT}
\title{Direct-INT}
\usage{
DINT(y, G, X, S, k = 3/8)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Snp by obs genotype matrix.}

\item{X}{Obs by feature covariate matrix.}

\item{S}{Obs by feature structure matrix.}

\item{k}{Offset applied during rank-normalization. See \code{\link{rankNormal}}.}
}
\value{
A numeric vector of p-values assessing the null hypothesis of no 
  genotypic effect. P-values are estimated using the Wald statistic, and 
  correspond to the rows of G.
}
\description{
Rank-normalizes the phenotype, then regresses the transformed phenotype on
genotype, covariates, and adjustments for population structure.
}
\examples{
# DINT against normal phenotype 
p = RNOmni::DINT(y=RNOmni::Y[,1],G=RNOmni::G[1:10,],X=RNOmni::X,S=RNOmni::S);
}
