% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitOLS}
\alias{fitOLS}
\title{Ordinary Least Squares}
\usage{
fitOLS(y, X)
}
\arguments{
\item{y}{Nx1 Numeric vector.}

\item{X}{NxP Numeric matrix.}
}
\value{
List containing the following:
\item{Beta}{Regression coefficient.}
\item{V}{Outcome variance.}
\item{Ibb}{Information matrix for beta.}
\item{Resid}{Outcome residuals.}
}
\description{
Fits the standard OLS model.
}
