% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{G}
\alias{G}
\title{Simulated Genotypes}
\format{A double matrix with 1000 rows and 1000 columns.
\describe{ 
   \item{s1-s1000}{Each column corresponds to genotype at 1000 loci (rows) for a single subject.}
}}
\usage{
G
}
\description{
Simulated genotypes for 1000 subjects (columns) at 1000 loci (rows).
Genotypes were simulated with use of hapgen2, and are based on the haplotype
structure of human chromosome one in the CEU population of the 1000 Genomes
Project.
}
\keyword{datasets}
