% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputCheck.R
\name{inCheck}
\alias{inCheck}
\title{Input Check}
\usage{
inCheck(y, G, X, S)
}
\arguments{
\item{y}{Numeric phenotype vector.}

\item{G}{Obs by snp genotype matrix.}

\item{X}{Obs by feature model matrix of covariates.}

\item{S}{Obs by feature model matrix of structure adjustments.}
}
\description{
Function to ensure the dimensions of inputs to association methods agree.
}
