% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OINT.R
\name{OmniP}
\alias{OmniP}
\title{Omnibus P-value.}
\usage{
OmniP(p)
}
\arguments{
\item{p}{Numeric vector of p-values.}
}
\value{
OINT p-value.
}
\description{
Obtains an omnibus p-value from a vector of potentially dependent p-values 
using the method of Cauchy combination. The p-values are converted to Cauchy
random deviates then averaged. The distribution of the average of these 
deviates is well-approximated by a Cauchy distribution in the tails. See
<https://doi.org/10.1080/01621459.2018.1554485>.
}
