% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OINT.R
\name{OmniP}
\alias{OmniP}
\title{Omnibus P-value}
\usage{
OmniP(u, r)
}
\arguments{
\item{u}{Test statistic.}

\item{r}{Correlation.}
}
\value{
Scalar.
}
\description{
Calculates the p-value for the maximum of two correlated, standard normal 
random variables.
}
