% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN.R
\name{RankNorm}
\alias{RankNorm}
\title{Rank-Normalize}
\usage{
RankNorm(u, k = 0.375, ties.method = "average")
}
\arguments{
\item{u}{Numeric vector.}

\item{k}{Offset. Defaults to (3/8), corresponding to the Blom transform.}

\item{ties.method}{Method of breaking ties, passed to \code{base::rank}.}
}
\value{
Numeric vector of rank normalized values.
}
\description{
Applies the rank-based inverse normal transform (INT) to a numeric vector.
The INT can be broken down into a two-step procedure. In the first, the
observations are transformed onto the probability scale using the empirical
cumulative distribution function (ECDF). In the second, the observations are
transformed onto the real line, as Z-scores, using the probit function.
}
\examples{
# Draw from chi-1 distribution
y <- stats::rchisq(n = 1e3, df = 1)
# Rank normalize
z <- RankNorm(y)
# Plot density of transformed measurement
plot(stats::density(z))
}
\seealso{
\itemize{
  \item Direct INT test \code{\link{DINT}}.
  \item Indirect INT test \code{\link{IINT}}.
  \item Omnibus INT test \code{\link{OINT}}.
}
}
