% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageHelp.R
\docType{package}
\name{RNOmni-help}
\alias{RNOmni-help}
\alias{RNOmni-help-package}
\title{RNOmni: Rank-Normal Omnibus Association Testing}
\description{
Implementation of genetic association tests that incorporate the rank-based inverse
normal transformation (INT) \code{\link{rankNorm}}. The direct-INT \code{\link{DINT}} test directly
transforms the outcome, whereas the indirect-INT \code{\link{IINT}} test forms residuals prior to transformation. 
The omnibus INT \code{\link{OINT}} test adaptively combines the D-INT and I-INT tests into a single
robust and statistically powerful procedure.
}
\author{
Zachary R. McCaw
}
