% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageHelp.R
\docType{package}
\name{RNOmni-help}
\alias{RNOmni-help}
\alias{RNOmni-help-package}
\title{RNOmni: Rank-Normal Omnibus Association Test}
\description{
Implementation of genetic association tests that utilize the rank-based inverse
normal transformation (INT). The main function is \code{\link{RNOmni}}, which performs
an omnibus, INT-based association test. The omnibus test combines the direct \code{\link{DINT}}
and indirect \code{\link{IINT}} INT-based tests. The transformation itself is described 
under \code{\link{rankNorm}}. See the vignette for additional details.
}
\author{
Zachary R. McCaw
}
