% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNOmni-package.R
\docType{package}
\name{RNOmni-package}
\alias{RNOmni}
\alias{RNOmni-package}
\title{RNOmni: Rank Normal Transformation Omnibus Test}
\description{
Inverse normal transformation (INT) based genetic association testing. These tests are recommend for continuous traits with non-normally distributed residuals. INT-based tests robustly control the type I error in settings where standard linear regression does not, as when the residual distribution exhibits excess skew or kurtosis. Moreover, INT-based tests outperform standard linear regression in terms of power. These tests may be classified into two types. In direct INT (D-INT), the phenotype is itself transformed. In indirect INT (I-INT), phenotypic residuals are transformed. The omnibus test (O-INT) adaptively combines D-INT and I-INT into a single robust and statistically powerful approach. See McCaw ZR, Lane JM, Saxena R, Redline S, Lin X. "Operating characteristics of the rank-based inverse normal transformation for quantitative trait analysis in genome-wide association studies" \doi{10.1111/biom.13214}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zrmacc/RNOmni}
}

}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
