% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAT.R
\name{ScoreStat}
\alias{ScoreStat}
\title{Score Statistics}
\usage{
ScoreStat(e, g, X, v)
}
\arguments{
\item{e}{Numeric residual vector.}

\item{g}{Genotype vector.}

\item{X}{Model matrix of covariates.}

\item{v}{Residual variance.}
}
\value{
Numeric vector containing the "score" statistic, standard error "se",
  "z", and "p" value.
}
\description{
Score Statistics
}
