% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{Y}
\alias{Y}
\title{Simulated Phenotypes}
\format{A numeric matrix with 1000 rows and 2 columns
\describe{
   \item{YN}{Normal phenotype.}
   \item{YL}{Log-normal phenotype.}
}}
\usage{
Y
}
\description{
Two independent phenotypes simulated under the null hypothesis of no
genotypic effect. A subject specific linear predictor was calculated based on
age, sex, pc1, and pc2. The normal phenotype was generated by adding N(0,1) noise
to the linear predictor. The log-normal phenotype was generated by adding N(0,1) noise,
then exponentiating.
}
\keyword{datasets}
